requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
ollama==0.1.7
biopython==1.83
openai==1.57.4
tiktoken==0.6.0
lxml==5.1.0
pandas==2.2.0
sentence-transformers>=3.0.0